DSA1
1. Develop a Program in C for the following:
A. Declare a calendar as an array of 7 elements (A dynamically Created array) to represent 7
days of a week. Each Element of the array is a structure having three fields. The first field
is the name of the Day (A dynamically allocated String), the second field is the date of the
Day (A integer), the third field is the description of the activity for a particular day (A
dynamically allocated String).
B. Write functions create(), read() and display(); to create the calendar, to read the data from
the keyboard and to print weeks activity details report on screen.