DSA2
2. Develop a Program in C for the following operations on Strings.
A. Read a main String (STR), a Pattern String (PAT) and a Replace String (REP)
B. Perform Pattern Matching Operation: Find and Replace all occurrences of PAT in STR
with REP if PAT exists in STR. Report suitable messages in case PAT does not exist in STR
Support the program with functions for each of the above operations. Don't use Built-in
functions.